pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "buildskill"
version = "0.1.0"
description = "从 GitHub 仓库分析并生成 Cursor Skill 文件"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "buildskill"}]
keywords = ["cursor", "skill", "github", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
]
dependencies = [
    "GitPython>=3.1.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
docx = ["python-docx"]

[project.scripts]
buildskill = "buildskill.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["buildskill*"]
